feat: migrate local password settings to authorization module and update related classes
